Add Role to Game class
