+++ src/game.js
@@ -1,9 +1,11 @@
 const fs = require('fs');
+const Role = require('../src/role.js');
 module.exports = class Game{
     constructor(creator, createdOn, players){
         this.Creator = creator;
         this.CreatedOn = createdOn;
         this.Players = players;
+        this.roles = this.loadRoles();
     }
 
     loadGames(){
@@ -17,4 +19,17 @@ module.exports = class Game{
 
         return gameList;
     }
+
+    loadRoles(){
+        let dataFile = fs.readFileSync('./data/roles.json', 'utf8');
+        let roleData = JSON.parse(dataFile);
+        let roleList = [];
+
+        roleData.forEach(element => {
+            roleList.push(new Role(element.roleID, element.title, element.description, element.image, element.color));
+        });
+
+        return roleList;
+    }
+
 }
